Moved  convert functions to data_utils, updated generate sentences to include report ids
