+++ sentence_classification/data_utils.py
@@ -1,5 +1,8 @@
 import csv
+import re
 import sys
+
+import nltk
 
 sys.path.append("..")  # Adds higher directory to python modules path.
 
@@ -23,9 +26,11 @@ def write_to_csv(sentence_file, sentence_tags):
         writer = csv.writer(fout, lineterminator='\n')
 
         # Write header 'sentence, labels'
-        writer.writerow(["sentence"] + ["processed sentence"] + ["diagnostic label"] + ["sentiment label"] + ["report id"])
+        writer.writerow(
+            ["sentence"] + ["processed sentence"] + ["diagnostic label"] + ["sentiment label"] + ["report id"])
         for row in sentence_tags:
-            writer.writerow([row.sentence] + [row.processed_sentence] + [row.diag_tag] + [row.sent_tag] + [row.report_id])
+            writer.writerow(
+                [row.sentence] + [row.processed_sentence] + [row.diag_tag] + [row.sent_tag] + [row.report_id])
 
 
 # Read in all sentences defined in the given file
@@ -46,37 +51,24 @@ def read_from_csv(sentence_file):
     # Return the read objects, but cut off the first row which was headers
     return data[1:]
 
-
-# Converts CSV in the original 3 column format into current version
-def read_from_csv_convert(sentenceFile):
-    data = []
-
-    with open(sentenceFile, 'rb') as fin:
-        reader = csv.reader(fin, delimiter=",")
-
-        for row in reader:
-            tmp = SentenceRecord(row[0])
-            tmp.processed_sentence = " ".join(preprocess.textPreprocess(row[0], removeNegationsFromSentences=False))
-            tmp.diag_tag = row[1]
-            tmp.sent_tag = row[2]
-            data.append(tmp)
-
-    # Return the read objects, but cut off the first row which was headers
-    return data[1:]
-
-
-# TODO rewrite to incorporate report IDs
 # Generate the unlabelled sentence data from the given datafile
-def generate_sentences(dataFiles):
-    sentence_tags = []
-    sentences = preprocess.getAllSentences(dataFiles)
+def generate_sentences(data_files):
+    data = []
+    tokenizer = nltk.data.load('tokenizers/punkt/english.pickle')
 
-    for sentence in sentences:
-        tmp = SentenceRecord(sentence)
-        tmp.processed_sentence = " ".join(preprocess.textPreprocess(sentence, removeNegationsFromSentences=False))
-        sentence_tags.append(tmp)
+    for file_name in data_files:
+        with open(file_name, 'rb') as file:
+            file.readline()  # skip header line
+            reader = csv.reader(file)
+            for row in reader:
+                for sentence in tokenizer.tokenize(row[1]):
+                    tmp = SentenceRecord(sentence)
+                    tmp.processed_sentence = " ".join(
+                        preprocess.textPreprocess(sentence, removeNegationsFromSentences=False))
+                    tmp.report_id = row[0]
+                    data.append(tmp)
 
-    return sentence_tags
+    return data
 
 
 def remove_duplicates(data):
@@ -94,12 +86,74 @@ def remove_duplicates(data):
 def generate_sentences_from_raw():
     confirm = raw_input("Are you sure you want to regenerate? (yes/no) ")
     if confirm == "yes":
-        types = ["Brains", "CTPA", "Plainab", "Pvab"]
-        files = []
-        for type in types:
-            files.append("../nlp_data/Cleaned" + type + "Full.csv")
+
+        files = ["./sentence_label_data/CleanedBrainsFull.csv",
+                 "./sentence_label_data/CleanedCTPAFull.csv",
+                 "./sentence_label_data/CleanedPlainabFull.csv",
+                 "./sentence_label_data/CleanedPvabFull.csv"]
 
         output_file = './sentence_label_data/sentences_ALL.csv'
         write_to_csv(output_file, remove_duplicates(generate_sentences(files)))
     else:
         print "Cancelled."
+
+
+def strip_labels(data):
+    for row in data:
+        row.diag_tag = ""
+        row.sent_tag = ""
+
+    return data
+
+
+def generate_sentence_id_dict():
+    sentence_files = ["./sentence_label_data/CleanedBrainsFull.csv",
+                      "./sentence_label_data/CleanedCTPAFull.csv",
+                      "./sentence_label_data/CleanedPlainabFull.csv",
+                      "./sentence_label_data/CleanedPvabFull.csv"]
+
+    # Dictionary <sentence, report ID>
+    sentence_ids = dict()
+
+    tokenizer = nltk.data.load('tokenizers/punkt/english.pickle')
+
+    for file_name in sentence_files:
+        with open(file_name, 'rb') as file:
+            file.readline()  # skip header line
+            reader = csv.reader(file)
+            for row in reader:
+                for sentence in tokenizer.tokenize(row[1]):
+                    if sentence not in sentence_ids:
+                        sentence_ids[sentence] = row[0]
+                    else:
+                        sentence_ids[sentence] = sentence_ids[sentence] + "," + row[0]
+
+    return sentence_ids
+
+
+def add_sentence_report_ids(data_file, sentence_id_file='./sentence_label_data/sentence_id_mappings.csv'):
+    csv.field_size_limit(sys.maxsize)
+
+    sentence_ids = dict(csv.reader(open(sentence_id_file, 'rb')))
+    labelled_data = []
+
+    with open(data_file, 'rb') as fin:
+        reader = csv.reader(fin, delimiter=",")
+
+        for row in reader:
+            tmp = SentenceRecord(row[0])
+            tmp.processed_sentence = row[1]
+            tmp.diag_tag = row[2]
+            tmp.sent_tag = row[3]
+            labelled_data.append(tmp)
+
+    # Return the read objects, but cut off the first row which was headers
+    labelled_data = labelled_data[1:]
+
+    # Set as first report ID corresponding to that sentence
+    for row in labelled_data:
+        query = row.sentence.replace('\\n', '')
+        query = query.replace('\n', '')
+        row.report_id = re.search("([0-9a-zA-Z]+)", sentence_ids[query]).group(0)
+
+    write_to_csv(data_file, labelled_data)
